+++ pkg/util/sysfs/manager/manager_linux.go
@@ -1,3 +1,6 @@
+//go:build linux
+// +build linux
+
 /*
 Copyright 2022 The Katalyst Authors.
