+++ pkg/util/sysfs/manager/sys.go
@@ -16,18 +16,6 @@ limitations under the License.
 
 package manager
 
-import (
-	"github.com/prometheus/procfs/sysfs"
-)
-
-func GetSystemCPUs() ([]sysfs.CPU, error) {
-	return GetSysFsManager().GetSystemCPUs()
-}
-
-func GetCPUTopology(string string) (*sysfs.CPUTopology, error) {
-	return GetSysFsManager().GetCPUTopology(string)
-}
-
 func GetNicRxQueueRPS(sysPath, nic string, queue int) (string, error) {
 	return GetSysFsManager().GetNicRxQueueRPS(sysPath, nic, queue)
 }
